+++ src/EightFaceMk2.cpp
@@ -82,7 +82,7 @@ struct EightFaceMk2Module : EightFaceMk2Base<NUM_PRESETS> {
 	SLOTCVMODE slotCvMode = SLOTCVMODE::TRIG_FWD;
 	SLOTCVMODE slotCvModeBak = SLOTCVMODE::OFF;
 	int slotCvModeDir = 1;
-	int slotCvModeAlt = 1;
+	int slotCvModeAlt = 0;
 	std::vector<int> slotCvModeShuffle;
 
 	std::default_random_engine randGen{(uint16_t)std::chrono::system_clock::now().time_since_epoch().count()};
@@ -93,6 +93,7 @@ struct EightFaceMk2Module : EightFaceMk2Base<NUM_PRESETS> {
 	dsp::SchmittTrigger slotC4Trigger;
 	dsp::SchmittTrigger resetTrigger;
 	dsp::Timer resetTimer;
+	bool resetFlag = false;
 
 	dsp::ClockDivider buttonDivider;
 	dsp::ClockDivider boundModulesDivider;
@@ -254,10 +255,17 @@ struct EightFaceMk2Module : EightFaceMk2Base<NUM_PRESETS> {
 		// Read mode
 		if (BASE::ctrlMode == CTRLMODE::READ) {
 			// RESET input
-			if (slotCvMode == SLOTCVMODE::TRIG_FWD || slotCvMode == SLOTCVMODE::TRIG_REV || slotCvMode == SLOTCVMODE::TRIG_PINGPONG) {
+			if (slotCvMode == SLOTCVMODE::TRIG_FWD ||
+				slotCvMode == SLOTCVMODE::TRIG_REV ||
+				slotCvMode == SLOTCVMODE::TRIG_PINGPONG ||
+				slotCvMode == SLOTCVMODE::TRIG_ALT ||
+				slotCvMode == SLOTCVMODE::TRIG_SHUFFLE) {
 				if (Module::inputs[INPUT_RESET].isConnected() && resetTrigger.process(Module::inputs[INPUT_RESET].getVoltage())) {
 					resetTimer.reset();
-					presetLoad(0);
+					resetFlag = true;
+					slotCvModeDir = 1;
+					slotCvModeAlt = 0;
+					slotCvModeShuffle.clear();
 				}
 			}
 
@@ -275,12 +283,14 @@ struct EightFaceMk2Module : EightFaceMk2Base<NUM_PRESETS> {
 						break;
 					case SLOTCVMODE::TRIG_FWD:
 						if (slotTrigger.process(Module::inputs[INPUT_CV].getVoltage())) {
-							presetLoad((preset + 1) % presetCount);
+							presetLoad(resetFlag ? 0 : ((preset + 1) % presetCount));
+							resetFlag = false;
 						}
 						break;
 					case SLOTCVMODE::TRIG_REV:
 						if (slotTrigger.process(Module::inputs[INPUT_CV].getVoltage())) {
-							presetLoad((preset - 1 + presetCount) % presetCount);
+							presetLoad(resetFlag ? (presetCount - 1) : ((preset - 1 + presetCount) % presetCount));
+							resetFlag = false;
 						}
 						break;
 					case SLOTCVMODE::TRIG_PINGPONG:
@@ -290,7 +300,8 @@ struct EightFaceMk2Module : EightFaceMk2Base<NUM_PRESETS> {
 								slotCvModeDir = -1;
 							if (n <= 0)
 								slotCvModeDir = 1;
-							presetLoad(n);
+							presetLoad(resetFlag ? 0 : n);
+							resetFlag = false;
 						}
 						break;
 					case SLOTCVMODE::TRIG_ALT:
@@ -304,7 +315,8 @@ struct EightFaceMk2Module : EightFaceMk2Base<NUM_PRESETS> {
 									slotCvModeDir = 1;
 								slotCvModeAlt = std::min(n, presetCount - 1);
 							}
-							presetLoad(n);
+							presetLoad(resetFlag ? 0 : n);
+							resetFlag = false;
 						}
 						break;
 					case SLOTCVMODE::TRIG_RANDOM:
@@ -338,6 +350,7 @@ struct EightFaceMk2Module : EightFaceMk2Base<NUM_PRESETS> {
 							int p = std::min(std::max(0, slotCvModeShuffle.back()), presetCount - 1);
 							slotCvModeShuffle.pop_back();
 							presetLoad(p);
+							resetFlag = false;
 						}
 						break;
 					case SLOTCVMODE::ARM: